change integer data type to long, cancel role requirement when user register and use mapper

+++ src/main/java/com/example/demo/dto/CreateUserDto.java
@@ -37,7 +37,7 @@ public class CreateUserDto {
     private String studentId;
 
     @NotNull(message = "Role ID is required")
-    private Integer roleId;
+    private Long roleId;
 
-    private Integer organizationalUnitId; // Optional, for staff members
+    private Long organizationalUnitId; // Optional, for staff members
 }
